Add basic HTTP client for Fleet.io API
In a full-fledged application, the API client would be separated
into multiple files with different methods. For now, one module
will do. We can split into successive commits.

The credentials to access the API are stored using Rails credentials.
In order to access them you need a copy of the master key.

Note that this version of the API client is not concerned with
pagination and is only fetching the first page of vehicles.
We will address this improvement in a separate commit.