Create a job to fetch vehicles from Fleet.io
The job will be scheduled to run every N minutes.
This is a crucial commit in the development of the application.
According to the instructions:

> There should be some sort of UI that allows someone to enter
> a VIN, and a way to view all the saved vehicles

Here all the approaches I considered before creating this job

1) Real-time API fetch

When the user types a VIN, we make a request to the API in real time.

There are two main issues.
Firstly, we are blocking the user in the browser while we wait for the
API to respond (which might be slow)
Secondly, there is no API endpoint to fetch a vehicle by VIN.
So we would have to fetch *all* the vehicles (through multiple pages)
until we find a matching VIN or get to the end.

2) Use external IDs

If there was a way to simply query the API with a VIN, then the
previous option would be viable. Fleet.io supports "External IDs"
which make this feasible.

Unfortunately, they can only be created
through the UI, and I have only access to the API token.

3) Fetch all vehicles once, then use webhooks for live updates

Another option is to fetch *all* the vehicles from the API just once,
when the app starts, and to store them in the database.
Then, whenever vehicles are created or updated, we can make use of
Fleet.io webhooks to be notified and update the database.

Unfortunately, webhooks can only be set for advanced accounts via UI.

4) Fetch all vehicles with a recurring job

Having considered all the options above and the tools I have, the
best solution is to fetch all the vehicles from the API, store them
in the database, and then fetch them again every N minutes, to make
sure the database is up-to-date.

To mitigate the lag between updates, we can provide users with a
way to explicitly start the job. That seems like a good compromise.

Having the list of vehicles in the local database makes for a very
snappy web experience, allowing to have nice UI components like
typeahead or autocomplete.