Let users "save" a vehicle
We add a new home-page that both shows the list of saved vehicles
and let the user type a VIN to save a new vehicle.

Rather than polluting the existing models and controllers for
Vehicle, we create a new controller and Active Model that only
focus on the act of saving (or bookmarking) a vehicle.

Note that the HTML in the view is functional but temporary and
will become nicer in successive commits.