Discard invalid vehicles from Fleet.io API
The Fleet.io API documentation does not clarify what to expect
in the vin/make/model attributes of each vehicle, and the sample
account provided for this exercises include some vehicles that
seem "erroneous", such as:

- "make": "",
- "model": "",
- "vin": "HTTPS://WWW.PIVOTALTRACKER.COM/STORY/SHOW/166170337",

Tests fail with these vehicles. My belief is that the data
is corrupted and we can assume that discarding these vehicles
is a good idea.