Set up job to fetch efficiency of vehicle
When saving a vehicle, launch a background process to fetch the
efficiency of the vehicle and update in the database.

Note that this commit simply sets up the job; the actual API call to
fleet.io will be dealt with separately.

Also note that, once the efficiency has been fetched, the user has
to hard-refresh the page to display it. Real-time update will be
added separately.