Add API client to fetch fuel entries
The API documentation about Fuel Entries displays many attributes.
It seems like the relevant ones to calculate total efficiency are
total gallons and total miles.

We can consider "Total gallons" as the sum of all the "us_gallons"
except for the last one (since we haven't technically 'used' that
fuel yet).

We can consider "Total miles" as the difference between the first
and the last reading of the meter (meter_entry : value).

I noticed that there is also an attribute called :usage_in_mi
but the data contained for some sample vehicles doesn't seem
reliable.