Add a convenience /fleet page (for admin use)
In order to know **which** VINs are valid, I have built a
convenience page at /fleet that displays all the vehicles
(saved or not).

This page is not part of the original specifications, but
it's useful to verify that the app works as intended.

As a bonus feature, I have added a "Refresh" button on the
fleet page in case we explicitly want to refresh the list
of vehicles from Fleet.io without waiting for the recurring
job to run.