Show fetched efficiency in real-time to the user
Before this commit, the user had to hard-refresh the page in order
to see the efficiency of a vehicle fetched from Fleet.io.

After this commit, the progress is displayed in real-time changing
from 'Calculating efficiency' to the actual efficiency, and falling
back to 'Not enough fuel entries' if it cannot be calculated.

Since all we needed is a simple pub/sub model with websockets,
I decided to go with Pusher rather than building a custom solution
with Action Cable.

A Pusher app has been set up and its credentials stored in the app.

Also note that the JavaScript to display the progress is very simple.
For this reason, it's vanilla JS embedded in the view. I will review
the overall approach about assets in a successive commit.