Make form faster by relying on Turbolinks
The form to save a vehicle doesn't need to hard-reload the page
once a vehicle has been saved. Turbolinks can make the experience
faster. The only gotcha is that Turbolinks doesn't work as well
with `render` as it does with `redirect_to`. We can quickly fix
this issue by including a separate gem.