Cleanup: load all assets from CDN
This commit is debatable and wouldn't be necessary in a full-fledged
app. However, for the scope of this project, we don't *need* the
full power of webpack and yarn because we can get all the assets
that we need from CDN:

- rails-ujs (JavaScript)
- turbolinks (JavaScript)
- pusher (JavaScript)
- bootstrap (CSS)

Therefore, in this specific case, it can be convenient to fully
rely on CDN to load these libraries. The advantage is that we don't
need to run webpack locally, we are not prone to local changes to
these libraries, and deployment will be faster (no precompilation).

This commit simply changes the reference to the assets.
The next commits will remove all the unnecessary files.