Demo only: show that efficiency is async
The Fleet.io API is so fast that the browser shows the efficiency
of a vehicle almost *immediately* after the vehicle has been saved.

Just for the purpose of demonstrating that efficiency is fetched
and measured in an async process, we add some forced delays to the
background process so we can clearly see when the message changes
from 'Calculating efficiency' to the calculated efficiency.